Introducing micro_states for create workflow
Introduce micro_states for the create workflow so as to
limit the granularity of the operation per task. The bigger
picture is to use the state transition diagram to make
decision which operation are permitted when two operations
try to modify the same volume object, helping in
serializing the requests. The micro_states table introduced
can be stored on a persistent store.

Part of blueprint cinder-state-enforcer

Part of blueprint valid-states-api

Co-Authored-By: Ivan Kolodyazhny <e0ne@e0ne.info>
Change-Id: Ica7680ac58b861ea19fbf990c5bd26764a497dc3